float amount services, fragment about, bug fixes

+++ app/src/main/java/com/bohdanserdyuk/KVPTPP/model/entity/ServiceData.kt
@@ -5,12 +5,12 @@ import android.arch.persistence.room.Ignore
 import android.arch.persistence.room.PrimaryKey
 
 @Entity(tableName = "serviceData")
-data class ServiceData(var title: String, var money: Int) {
+data class ServiceData(var title: String, var money: Float) {
 
     @PrimaryKey(autoGenerate = true)
     var id: Int = 0
 
-   @Ignore constructor(id: Int, title: String, money: Int) : this(title, money) {
+   @Ignore constructor(id: Int, title: String, money: Float) : this(title, money) {
         this.id = id
     }
 }
